requirements.txt
flask
requests
python-dotenv
beautifulsoup4
playwright
